+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/impl/DeveloperDAOImpl.java
@@ -4,8 +4,6 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-import javax.persistence.NoResultException;
-import javax.persistence.NonUniqueResultException;
 import javax.persistence.Query;
 
 import org.apache.logging.log4j.LogManager;
@@ -18,19 +16,17 @@ import org.springframework.util.StringUtils;
 import gov.healthit.chpl.auth.Util;
 import gov.healthit.chpl.dao.AddressDAO;
 import gov.healthit.chpl.dao.ContactDAO;
+import gov.healthit.chpl.dao.DeveloperDAO;
 import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
-import gov.healthit.chpl.dao.DeveloperDAO;
 import gov.healthit.chpl.dto.CertifiedProductDTO;
-import gov.healthit.chpl.dto.ContactDTO;
 import gov.healthit.chpl.dto.DeveloperACBMapDTO;
 import gov.healthit.chpl.dto.DeveloperDTO;
 import gov.healthit.chpl.entity.AttestationType;
-import gov.healthit.chpl.entity.CertifiedProductEntity;
 import gov.healthit.chpl.entity.ContactEntity;
 import gov.healthit.chpl.entity.DeveloperACBMapEntity;
+import gov.healthit.chpl.entity.DeveloperACBTransparencyMapEntity;
 import gov.healthit.chpl.entity.DeveloperEntity;
-import gov.healthit.chpl.entity.ProductEntity;
 
 @Repository("developerDAO")
 public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
@@ -38,11 +34,11 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 	private static final Logger logger = LogManager.getLogger(DeveloperDAOImpl.class);
 	@Autowired AddressDAO addressDao;
 	@Autowired ContactDAO contactDao;
-	
+
 	@Override
 	@Transactional
 	public DeveloperDTO create(DeveloperDTO dto) throws EntityCreationException, EntityRetrievalException {
-		
+
 		DeveloperEntity entity = null;
 		try {
 			if (dto.getId() != null){
@@ -51,10 +47,10 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		} catch (EntityRetrievalException e) {
 			throw new EntityCreationException(e);
 		}
-		
+
 		if (entity != null) {
 			throw new EntityCreationException("An entity with this ID already exists.");
-		} else {			
+		} else {
 			entity = new DeveloperEntity();
 
 			if(dto.getAddress() != null)
@@ -62,7 +58,6 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 				entity.setAddress(addressDao.mergeAddress(dto.getAddress()));
 			}
 			if(dto.getContact() != null) {
-				ContactEntity contact = null;
 				if(dto.getContact().getId() != null) {
 					Query query = entityManager.createQuery("from ContactEntity a where (NOT deleted = true) AND (contact_id = :entityid) ", ContactEntity.class );
 					query.setParameter("entityid", dto.getContact().getId());
@@ -74,37 +69,37 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 					entity.setContact(contactDao.create(dto.getContact()));
 				}
 			}
-			
+
 			entity.setName(dto.getName());
 			entity.setWebsite(dto.getWebsite());
-			
+
 			if(dto.getDeleted() != null) {
 				entity.setDeleted(dto.getDeleted());
 			} else {
 				entity.setDeleted(false);
 			}
-			
+
 			if(dto.getLastModifiedUser() != null) {
 				entity.setLastModifiedUser(dto.getLastModifiedUser());
 			} else {
 				entity.setLastModifiedUser(Util.getCurrentUser().getId());
-			}		
-			
+			}
+
 			if(dto.getLastModifiedDate() != null) {
 				entity.setLastModifiedDate(dto.getLastModifiedDate());
 			} else {
 				entity.setLastModifiedDate(new Date());
 			}
-			
+
 			if(dto.getCreationDate() != null) {
 				entity.setCreationDate(dto.getCreationDate());
 			} else {
 				entity.setCreationDate(new Date());
 			}
-			
+
 			create(entity);
 			return new DeveloperDTO(entity);
-		}	
+		}
 	}
 
 	@Override
@@ -121,16 +116,16 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		entityManager.flush();
 		return new DeveloperACBMapDTO(mapping);
 	}
-	
+
 	@Override
 	@Transactional
 	public DeveloperEntity update(DeveloperDTO dto) throws EntityRetrievalException {
 		DeveloperEntity entity = this.getEntityById(dto.getId());
-		
+
 		if(entity == null) {
 			throw new EntityRetrievalException("Entity with id " + dto.getId() + " does not exist");
 		}
-		
+
 		if(dto.getAddress() != null)
 		{
 			try {
@@ -142,9 +137,8 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		} else {
 			entity.setAddress(null);
 		}
-		
+
 		if(dto.getContact() != null) {
-			ContactEntity contact = null;
 			if(dto.getContact().getId() != null) {
 				Query query = entityManager.createQuery("from ContactEntity a where (NOT deleted = true) AND (contact_id = :entityid) ", ContactEntity.class );
 				query.setParameter("entityid", dto.getContact().getId());
@@ -171,28 +165,28 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		} else {
 			entity.setContact(null);
 		}
-		
+
 		entity.setWebsite(dto.getWebsite());
 		if(dto.getName() != null) {
 			entity.setName(dto.getName());
 		}
-		
+
 		if(dto.getDeleted() != null) {
 			entity.setDeleted(dto.getDeleted());
 		}
-		
+
 		if(dto.getLastModifiedUser() != null) {
 			entity.setLastModifiedUser(dto.getLastModifiedUser());
 		} else {
 			entity.setLastModifiedUser(Util.getCurrentUser().getId());
-		}		
-		
+		}
+
 		if(dto.getLastModifiedDate() != null) {
 			entity.setLastModifiedDate(dto.getLastModifiedDate());
 		} else {
 			entity.setLastModifiedDate(new Date());
 		}
-			
+
 		update(entity);
 		return entity;
 	}
@@ -203,7 +197,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		if(mapping == null) {
 			return null;
 		}
-		
+
 		mapping.setTransparencyAttestation(AttestationType.getValue(dto.getTransparencyAttestation()));
 		mapping.setLastModifiedDate(new Date());
 		mapping.setLastModifiedUser(Util.getCurrentUser().getId());
@@ -211,12 +205,12 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		entityManager.flush();
 		return new DeveloperACBMapDTO(mapping);
 	}
-	
+
 	@Override
 	@Transactional
 	public void delete(Long id) throws EntityRetrievalException {
 		DeveloperEntity toDelete = getEntityById(id);
-		
+
 		if(toDelete != null) {
 			toDelete.setDeleted(true);
 			toDelete.setLastModifiedDate(new Date());
@@ -224,7 +218,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 			update(toDelete);
 		}
 	}
-	
+
 	@Override
 	public void deleteTransparencyMapping(Long developerId, Long acbId) {
 		DeveloperACBMapEntity toDelete = getTransparencyMappingEntity(developerId, acbId);
@@ -236,13 +230,13 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 			entityManager.flush();
 		}
 	}
-	
+
 	@Override
 	public List<DeveloperDTO> findAll() {
-		
+
 		List<DeveloperEntity> entities = getAllEntities();
 		List<DeveloperDTO> dtos = new ArrayList<>();
-		
+
 		for (DeveloperEntity entity : entities) {
 			DeveloperDTO dto = new DeveloperDTO(entity);
 			dtos.add(dto);
@@ -258,10 +252,22 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		}
 		return new DeveloperACBMapDTO(mapping);
 	}
-	
+
+	@Override
+	public List<DeveloperACBMapDTO> getAllTransparencyMappings() {
+		List<DeveloperACBTransparencyMapEntity> entities = getTransparencyMappingEntities();
+        List<DeveloperACBMapDTO> dtos = new ArrayList<>();
+
+        for (DeveloperACBTransparencyMapEntity entity : entities) {
+            DeveloperACBMapDTO dto = new DeveloperACBMapDTO(entity);
+            dtos.add(dto);
+        }
+        return dtos;
+	}
+
 	@Override
 	public DeveloperDTO getById(Long id) throws EntityRetrievalException {
-		
+
 		DeveloperEntity entity = getEntityById(id);
 		DeveloperDTO dto = null;
 		if(entity != null) {
@@ -269,7 +275,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		}
 		return dto;
 	}
-	
+
 	@Override
 	public DeveloperDTO getByName(String name) {
 		DeveloperEntity entity = getEntityByName(name);
@@ -279,7 +285,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		}
 		return dto;
 	}
-	
+
 	@Override
 	public DeveloperDTO getByCode(String code) {
 		DeveloperEntity entity = getEntityByCode(code);
@@ -289,17 +295,17 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		}
 		return dto;
 	}
-	
+
 	public DeveloperDTO getByCertifiedProduct(CertifiedProductDTO cpDto) throws EntityRetrievalException {
 		if(cpDto == null || cpDto.getProductVersionId() == null) {
 			throw new EntityRetrievalException("Version ID cannot be null!");
 		}
 		Query getDeveloperByVersionIdQuery = entityManager.createQuery(
 				"FROM ProductVersionEntity pve,"
-				+ "ProductEntity pe, DeveloperEntity ve " 
+				+ "ProductEntity pe, DeveloperEntity ve "
 				+ "WHERE (NOT pve.deleted = true) "
 				+ "AND pve.id = :versionId "
-				+ "AND pve.productId = pe.id " 
+				+ "AND pve.productId = pe.id "
 				+ "AND ve.id = pe.developerId ", DeveloperEntity.class);
 		getDeveloperByVersionIdQuery.setParameter("versionid", cpDto.getProductVersionId());
 		Object result = getDeveloperByVersionIdQuery.getSingleResult();
@@ -309,33 +315,31 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		DeveloperEntity ve = (DeveloperEntity)result;
 		return new DeveloperDTO(ve);
 	}
-	
+
 	private void create(DeveloperEntity entity) {
-		
 		entityManager.persist(entity);
 		entityManager.flush();
 	}
-	
+
 	private void update(DeveloperEntity entity) {
-		
-		entityManager.merge(entity);	
+		entityManager.merge(entity);
 		entityManager.flush();
 	}
-	
+
 	private List<DeveloperEntity> getAllEntities() {
 		List<DeveloperEntity> result = entityManager.createQuery( "SELECT v from "
 				+ "DeveloperEntity v "
 				+ "LEFT OUTER JOIN FETCH v.address "
 				+ "LEFT OUTER JOIN FETCH v.contact "
+				+ "LEFT OUTER JOIN FETCH v.developerCertificationStatuses "
 				+ "where (NOT v.deleted = true)", DeveloperEntity.class).getResultList();
 		return result;
-		
 	}
 
 	private DeveloperEntity getEntityById(Long id) throws EntityRetrievalException {
-		
+
 		DeveloperEntity entity = null;
-			
+
 		Query query = entityManager.createQuery( "SELECT v from "
 				+ "DeveloperEntity v "
 				+ "LEFT OUTER JOIN FETCH v.address "
@@ -343,7 +347,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 				+ "where (NOT v.deleted = true) AND (vendor_id = :entityid) ", DeveloperEntity.class );
 		query.setParameter("entityid", id);
 		List<DeveloperEntity> result = query.getResultList();
-		
+
 		if (result.size() > 1){
 			throw new EntityRetrievalException("Data error. Duplicate developer id in database.");
 		} else if(result.size() == 1) {
@@ -352,11 +356,11 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 
 		return entity;
 	}
-	
+
 	private DeveloperEntity getEntityByName(String name) {
-		
+
 		DeveloperEntity entity = null;
-			
+
 		Query query = entityManager.createQuery( "SELECT v from "
 				+ "DeveloperEntity v "
 				+ "LEFT OUTER JOIN FETCH v.address "
@@ -364,18 +368,18 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 				+ "where (NOT v.deleted = true) AND (v.name = :name) ", DeveloperEntity.class );
 		query.setParameter("name", name);
 		List<DeveloperEntity> result = query.getResultList();
-		
+
 		if(result.size() > 0) {
 			entity = result.get(0);
 		}
 
 		return entity;
 	}
-	
+
 	private DeveloperEntity getEntityByCode(String code) {
-		
+
 		DeveloperEntity entity = null;
-			
+
 		Query query = entityManager.createQuery( "SELECT v from "
 				+ "DeveloperEntity v "
 				+ "LEFT OUTER JOIN FETCH v.address "
@@ -383,7 +387,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 				+ "where (NOT v.deleted = true) AND (v.developerCode = :code) ", DeveloperEntity.class );
 		query.setParameter("code", code);
 		List<DeveloperEntity> result = query.getResultList();
-		
+
 		if(result.size() > 0) {
 			entity = result.get(0);
 		}
@@ -399,11 +403,16 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 				+ "AND map.certificationBodyId = :acbId", DeveloperACBMapEntity.class);
 		query.setParameter("developerId", developerId);
 		query.setParameter("acbId", acbId);
-		
+
 		List<DeveloperACBMapEntity> results = query.getResultList();
 		if(results != null && results.size() > 0) {
 			return results.get(0);
 		}
 		return null;
+	}
+
+	private List<DeveloperACBTransparencyMapEntity> getTransparencyMappingEntities() {
+		List<DeveloperACBTransparencyMapEntity> result = entityManager.createQuery( "FROM DeveloperACBTransparencyMapEntity",DeveloperACBTransparencyMapEntity.class).getResultList();
+		return result;
 	}
 }